+++ scripts/legacyMacros.js
@@ -649,6 +649,7 @@ export {wingsOfFlying, wingsOfFlyingEffect, wingsOfFlyingRecharge} from './macro
 export {wrapsOfDyamak, wrapsOfDyamak0, wrapsOfDyamak1, wrapsOfDyamak2, crimsonMist} from './macros/2014/items/trinket/wrapsOfDyamak.js';
 export {robeOfUsefulItems, robeOfUsefulItemsSetupPatch, robeOfUsefulItemsPatch} from './macros/2014/items/clothing/robeOfUsefulItems.js';
 export {ruinsWakeD, ruinsWakeA, ruinsWakeE} from './macros/2014/items/weapons/ruinsWake.js';
+export {charlatansDie} from './macros/2014/items/trinket/charlatansDie.js';
 //BG3 Weapon Actions
 export {backbreaker, braceMelee, braceMeleeDamage, cleave, concussiveSmash, maimingStrike, disarmingStrike, flourish, heartstopper, lacerate, piercingStrike, piercingStrikeEffect, pommelStrike, prepare, rush, tenacity, topple, weakeningStrike, braceRanged, braceRangedDamage, hamstringShot, mobileShot, piercingShot} from './macros/2014/homebrew/bg3WeaponActions.js';
 //Other